Start restructure project
