+++ setup.cfg
@@ -1,0 +1,21 @@
+[metadata]
+name = python_velbustcp
+description = Bridge a Velbus installation with TCP
+author = Velleman NV
+
+[options]
+packages = find_namespace:
+install_requires =
+    pyserial==3.4
+    pytz==2019.3
+    tzlocal==2.0.0
+python_requires = >=3.6
+package_dir =
+    =src
+zip_safe = no
+
+[options.package_data]
+python-velbustcp = py.typed
+
+[options.packages.find]
+where = src
