+++ src/velbustcp/lib/consts.py
